+++ src/site/markdown/overview/changelog.md
@@ -29,6 +29,7 @@
 
 *   Java
     *   [#1530](https://sourceforge.net/p/pmd/bugs/1530/): \[java] Parser exception on Java code
+    *   [#1490](https://sourceforge.net/p/pmd/bugs/1490/): \[java] PMD Error while processing - NullPointerException
 *   java-basic/BooleanInstantiation
     *   [#1533](https://sourceforge.net/p/pmd/bugs/1533/): \[java] BooleanInstantiation: ClassCastException with Annotation
 *   java-comments